+++ src-docs/src/views/datagrid/advanced/datagrid_advanced_example.js
@@ -1,7 +1,13 @@
 import React from 'react';
 
 import { GuideSectionTypes } from '../../../components';
-import { EuiCode, EuiSpacer, EuiCallOut } from '../../../../../src/components';
+import {
+  EuiCode,
+  EuiSpacer,
+  EuiCallOut,
+  EuiTitle,
+  EuiLink,
+} from '../../../../../src/components';
 
 import { EuiDataGridRefProps } from '!!prop-loader!../../../../../src/components/datagrid/data_grid_types';
 
@@ -81,8 +87,6 @@ export const DataGridAdvancedExample = {
             </li>
           </ul>
 
-          <EuiSpacer size="s" />
-
           <EuiCallOut title="Handling cell location">
             When using <EuiCode>setFocusedCell</EuiCode> or{' '}
             <EuiCode>openCellPopover</EuiCode>, keep in mind:
@@ -104,11 +108,74 @@ export const DataGridAdvancedExample = {
             </ul>
           </EuiCallOut>
 
+          <EuiSpacer size="s" />
+
+          <EuiTitle>
+            <h3>react-window methods</h3>
+          </EuiTitle>
+          <p>
+            <EuiCode>EuiDataGrid</EuiCode> also exposes several underlying
+            methods from{' '}
+            <EuiLink
+              href="https://react-window.vercel.app/#/api/VariableSizeGrid"
+              target="_blank"
+            >
+              react-window&apos;s <EuiCode>VariableSizeGrid</EuiCode> imperative
+              API
+            </EuiLink>{' '}
+            via its <EuiCode>ref</EuiCode>:
+          </p>
+          <ul>
+            <li>
+              <p>
+                <EuiCode>
+                  scrollTo({'{ scrollLeft: number, scrollTop: number }'})
+                </EuiCode>{' '}
+                - scrolls the grid to the specified horizontal and vertical
+                pixel offsets.
+              </p>
+            </li>
+            <li>
+              <p>
+                <EuiCode>
+                  scrollToItem(
+                  {
+                    '{ align: string = "auto", columnIndex?: number, rowIndex?: number }'
+                  }
+                  )
+                </EuiCode>{' '}
+                - scrolls the grid to the specified row and columns indices
+              </p>
+            </li>
+          </ul>
+          <EuiCallOut title="react-window vs. EUI">
+            <p>
+              Unlike EUI&apos;s ref APIs, <EuiCode>rowIndex</EuiCode> here
+              refers to the <strong>visible</strong> <EuiCode>rowIndex</EuiCode>{' '}
+              when passed to a method of a native{' '}
+              <EuiCode>react-window</EuiCode> API.
+            </p>
+            <p>
+              For example:{' '}
+              <EuiCode>
+                {'scrollToItem({ rowIndex: 50, columnIndex: 0 })'}
+              </EuiCode>{' '}
+              will always scroll to 51st visible row on the currently visible
+              page, regardless of the content in the cell. In contrast,{' '}
+              <EuiCode>
+                {'setFocusedCell({ rowIndex: 50, colIndex: 0 })'}
+              </EuiCode>{' '}
+              will scroll to the 51st row in your data, which may not be the
+              51st visible row in the grid if it is paginated or sorted.
+            </p>
+          </EuiCallOut>
+
           <EuiSpacer />
 
           <p>
             The below example shows how to use the internal APIs for a data grid
-            that opens a modal via cell actions.
+            that opens a modal via cell actions, that scroll to specific cells,
+            and that can be put into full-screen mode.
           </p>
         </>
       ),